// SIC/SIC Admin/src/types/seasonalTopics.ts
export type AddEditSeasonalActivityData = {
	onSubmit: () => void;
	onClose: () => void;
	activityUuid: string;
	url: string;
};